# [core] improve Color & Clamp performance