--- a/Xamarin.Forms.Core/NumericExtensions.cs
+++ b/Xamarin.Forms.Core/NumericExtensions.cs
@@ -1,20 +1,48 @@
 using System;
 using System.ComponentModel;
+using System.Runtime.CompilerServices;
 
 namespace Xamarin.Forms.Internals
 {
 	[EditorBrowsable(EditorBrowsableState.Never)]
 	public static class NumericExtensions
 	{
-
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static double Clamp(this double self, double min, double max)
 		{
-			return Math.Min(max, Math.Max(self, min));
+			if (max < min)
+			{
+				return max;
+			}
+			else if (self < min)
+			{
+				return min;
+			}
+			else if (self > max)
+			{
+				return max;
+			}
+
+			return self;
 		}
 
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static int Clamp(this int self, int min, int max)
 		{
-			return Math.Min(max, Math.Max(self, min));
+			if (max < min)
+			{
+				return max;
+			}
+			else if (self < min)
+			{
+				return min;
+			}
+			else if (self > max)
+			{
+				return max;
+			}
+
+			return self;
 		}
 	}
 }