@jonathanpeppers 2019-12-18
Yeah it's called 4 times for rgba colors: https://github.com/xamarin/Xamarin.Forms/blob/9d019b90ea76e7d85456523c1a848109c1d87cb3/Xamarin.Forms.Core/Color.cs#L101-L104

Then there is a list of ~142 colors: https://github.com/xamarin/Xamarin.Forms/blob/9d019b90ea76e7d85456523c1a848109c1d87cb3/Xamarin.Forms.Core/Color.cs#L421-L564

I have another idea that does the math for these at build time: https://github.com/xamarin/Xamarin.Forms/compare/master...jonathanpeppers:colorgen

Still working on that one ☝️