@StephaneDelcroix 2019-12-18
@jonathanpeppers that looks cool, but a lot of work for very little benefits.

we could:
- have a private ctor that bypass the Clamp,
- or precompute each value and directly assign the components (precompute the HSL components too)