@jonathanpeppers 2019-12-18
@StephaneDelcroix instead of generating the `cs` file like this: https://github.com/xamarin/Xamarin.Forms/compare/master...jonathanpeppers:colorgen

Do these color change often? As I could just run this "script" once, and we just commit the list of colors with precomputed `float` values.