@StephaneDelcroix 2019-12-18
they never change. committing the result (with the ld values as comment) would be good