@jonathanpeppers 2019-12-18
That idea didn't actually pan out: https://github.com/xamarin/Xamarin.Forms/compare/4.4.0...jonathanpeppers:precompute-color-floats

My guess is the JIT time + actual larger data is slower than the math that happens here.