@StephaneDelcroix 2019-12-18
but we could still bypass the Clamp for predefined colors