@jonathanpeppers 2019-12-18
> but we could still bypass the Clamp for predefined colors

Yeah, let me try defining the three floats for RGB and still do the math for HSL.