@ahmedalejo 2020-01-06
Hi @jonathanpeppers could you benchmark the following.
Most of the existent colors aren´t used in a common app´s life-cycle (at least not at startup)

```csharp
public partial struct Color
{
	private static readonly Color aliceBlue;
	public static Color AliceBlue => aliceBlue ??= new Color(240, 248, 255);
...
```