@jonathanpeppers 2020-01-07
@ahmedalejo can you open a PR? (or link to a diff) I can time the difference.

The main issue is the API change, going from a field to a property can break existing code.