@StephaneDelcroix 2020-01-07
> going from a field to a property can break existing code.

it WILL break ABI. We can't do that