@ahmedalejo 2020-01-07
Hi @StephaneDelcroix 
- i get your point. But "ABI" really meant "API" and not "Android ABI" right?

Hi @jonathanpeppers here a pull request
- https://github.com/xamarin/Xamarin.Forms/pull/9109